Implement temporary fix for constantly neeeding to log in
The cookies sent back from the server only seem to be persisted for the
length of the session, which means as soon as the browser is closed, the
session is lost. As a temporary workaround, I've re-implemented basic
auth with client-side refreshing of the cookie to keep it authenticated.
This is obviously not a good decision but it buys be time to investigate
backend solutions.